% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{data}
\name{assay}
\alias{assay}
\title{assay data for 1536 samples and 92 snps}
\format{
A data frame with 1536 rows and 94 variables. The variables
are as follows:
\describe{
  \item{Well}{The Well number in each plate}
  \item{Sample.Name}{The ID of each sample}
  \item{snp locis}{The ID of each snp loci in the Well plate}
}
}
\source{
unknown.
}
\usage{
assay
}
\description{
A dataset containing the information of 1536 samples
and their 92 snp locis data.
}
\keyword{datasets}
