% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.assay.to.ped.R
\name{from.assay.to.ped}
\alias{from.assay.to.ped}
\title{transform the output of the function \code{read.assay.file} into data.ped.}
\usage{
from.assay.to.ped(
  assay,
  snp.info,
  sample.info = NA,
  is.sample.info = F,
  col.names.convert = T,
  fam.id = F,
  is.num = T,
  write.file = F
)
}
\arguments{
\item{assay}{output of the function \code{read.assay.file}.}

\item{snp.info}{a data frame containing the SNP information include the assay name or rs id, the ref allele, alter allele of each SNP.}

\item{sample.info}{a data frame containing the sample information include the id, Sex and phenotype of each sample.}

\item{is.sample.info}{logical. If \code{TRUE}, the sample.info is required and the funtion will merge the sample information to the final output.}

\item{col.names.convert}{logical. if \code{TRUE}, convert the colnames of assay to the rs id of each SNP, such as: rs001.}

\item{fam.id}{logical. if \code{TRUE}, the ped data will contain the family data of the samples.}

\item{is.num}{logical. if \code{TRUE}, the loci data in ped data will be like "1 1", "1 2" or "2 2". if \code{FALSE}, the loci data in ped data will be like "A A","A B" or "B B".}

\item{write.file}{logical. if \code{TRUE}, the function will write ped data down.}
}
\value{
a data frame with the col is sample and the row is snp.loci:
\item{ped}{function output}
}
\description{
\code{from.assay.to.ped} transform the output of function \code{read.assay.file} into data.ped 
then you can save the data.ped as file.ped. before using this function, you need to prepare the 
information of SNP and samples.
}
\details{
This function use the output of the function \code{read.assay.file}.
if the snp.info and sample.info data are available, you should input
them, too. The output of this function is a data frame in ped format. 
The first 6 col of output are about sample information. And the other 
col are the snp from the assay files.
}
\examples{
data.ped <- from.assay.to.ped(data.total.1536, snp.info = snp.info, sample.info = sample.info, is.sample.info = T)


}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
