% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.gene.causal.R
\name{gene.gene.causal}
\alias{gene.gene.causal}
\title{calculate the causal relationship between the interacting gene pairs}
\usage{
gene.gene.causal(
  data.ggiud,
  n.times = 1000,
  sample.size = 100,
  ref = c("Gaussian", "uniform"),
  est = c("entropy", "intergral"),
  interaction.p = 0.05,
  fex = T,
  pos.use = T
)
}
\arguments{
\item{n.times}{the times of the permutation.}

\item{sample.size}{the sample.size of each time of permutation}

\item{ref}{(optional) a character string matching one of the following items: Gaussian and uniform. the data distribution of your data in function \code{igci}}

\item{est}{(optional) a character string matching one of the following items: entropy and intergral. the estimator used in function \code{igci}}

\item{interaction.p}{the p.value threshold of the interacting relationship, default: 0.05.}

\item{fex}{logical. if \code{TRUE}, using the function \code{fex} to reduce the dimensions of SnpMatrix.}

\item{data.gguid}{the ouput of the function \code{gene.gene.interaction}, a list of the class GGIUD.}
}
\value{
a GGC.data list contain 6 item:
\item{SnpMatrix}{function output}
\item{gene.info}{gene info}
\item{sample.info}{sample info}
\item{gg.causal}{gene gene causal table}
\item{gene.table}{gene table}
\item{gg.f}{gene f.value table}
}
\description{
\code{gene.gene.causal} calculate the causal relationship between the interacting gene pairs, although you can calculate the causal relationship between any gene pairs, that doesn't make senses.
}
\details{
After the function \code{gene.gene.interaction} calculates the 
interaction between genes, the causal relationship between gene 
pairs with obvious interaction is inferred. When f<0, gene 1 leads 
to gene 2, and f>0 is the opposite.
}
\examples{
temp.ggc <- gene.gene.causal(temp.ggiud, n.times = 1000, sample.size = 700, fex = TRUE)

}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
