% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.gene.interaction.R
\name{gene.gene.interaction}
\alias{gene.gene.interaction}
\title{calculate the gene-gene interaction p.val and s.val and return the result with the input data.}
\usage{
gene.gene.interaction(
  data,
  n.times = 1000,
  selected.gene.list = NULL,
  gene.level = c("gene.level", "default", "minP")
)
}
\arguments{
\item{data}{output of the function \code{impute.res}.}

\item{n.times}{the times of the permutation.}

\item{selected.gene.list}{a vector.you can selected part of genes in the gene.info.}

\item{gene.level}{a character string matching one of the following item: gene.level, default, minP. more information is available in the details.}
}
\value{
a list of class GGIUD contain 7 items:
\item{SnpMatrix}{function output}
\item{gene.info}{genetic information of each snp loci}
\item{sample.info}{sample information contain Sex and phenotype}
\item{GGI.ud}{gene-gene interaction table, you can plot a network by using this table and the package\code{igrah}}
\item{GENE.table}{the selected gene list and the number of snps on each gene}
\item{P.matrix}{p value matrix}
\item{S.matrix}{delta H matrix}
}
\description{
\code{gene.gene.interaction} calculate the gene-gene interaction p.val and s.val based on the method GBIGM.
}
\details{
This function use the output of the function \code{impute.res}. The 
class of the input data should be GGIGCI.data. The interaction calculated 
by this function is based on the information gained method, so the interaction 
can be non-linear. Larger n.times you set, more time this function run. In order 
to reduce the time usage, you can set the param:gene.level as minP to make this
function run faster. And when your input data size is too big to get a desired 
result, you can also try to set gene.level as minP. On the other hand, if the number
of SNPs on your genes are too few to present the genes, you should set gene.level as 
gene.level to prevent the wrong result.
}
\examples{
temp.ggiud <- gene.gene.interaction(temp.im, n.times = 1000, selected.gene.list = c("DRD1", "DRD2", "DRD3"))


}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
