% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRACI.R
\docType{package}
\name{GRACI-package}
\alias{GRACI-package}
\alias{GRACI}
\title{Functions for finding the causality between genes by using the SnpMatrix data and plotting the relationship network.}
\description{
GRACI provides tools for calculating and visualizing
causality relationship among genes.
}
