% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.res.R
\name{impute.res}
\alias{impute.res}
\title{impute the SnpMatrix}
\usage{
impute.res(res, removed = c("snps", "sample", "none"), silence = TRUE)
}
\arguments{
\item{res}{output of the function \code{read.ped.to.snpMatrix}.}

\item{removed}{(optional) a character string matching one of the following items: snps, sample, none. Describes the action taken in case of remaining missing values.}

\item{silence}{logical (default:TRUE). If \code{TRUE}, this function won't print the detail of imputation.}
}
\value{
a list of class GGIGCI.data with 3 items:
\item{SnpMatrix}{SnpMatrix contain each sample's each snp's genotype}
\item{gene.info}{genetic information of each snp loci}
\item{sample.info}{sample information contain Sex and phenotype}
}
\description{
\code{impute.res} impute the SnpMatrix and remove the remain missing value in the matrix.
}
\details{
This function mimics a Leave-One-Out process where missing SNPs are 
imputed for an individual based on a model trained on all other individuals.
This function use the output of the function \code{read.ped.to.snpMatrix}.
}
\examples{
temp.im <- impute.res(res, removed = "sample")


}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
