% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GGC.data.R
\name{plot.GGC.data}
\alias{plot.GGC.data}
\title{the plot function for the Object of class GGC.data}
\usage{
\method{plot}{GGC.data}(
  ggc,
  color.method = c("snp", "out"),
  color.threshold = 5,
  shape.method = c("snp", "out"),
  shape.threshold = 5
)
}
\arguments{
\item{ggc}{output of the function \code{gene.gene.causal}.}

\item{color.method}{(optional) a character string matching one of the following items: snp and out.}

\item{color.threshold}{the threshold to define the color of each node in the network.}

\item{shape.method}{(optional) a character string matching one of the following items: snp and out.}

\item{shape.threshold}{the threshold to define the shape of each node in the network.}
}
\value{
a network plot contain all the gene and the relationship between them.
}
\description{
\code{plot} For the object of class GGC.data, plot the object as a network.
}
\details{
This function use the output of the function \code{gene.gene.causal}. 
The output image of this function is a network. Every node in the network
is a gene, and every edge in the network is a causal relationship between 
two genes. And the node can be plot in different color or shape.
}
\examples{
plot(temp.ggc, color.method = "snp", shape.method = "out")


}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
