% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ped.to.snpMatrix.R
\name{read.ped.to.snpMatrix}
\alias{read.ped.to.snpMatrix}
\title{read the ped file and output a list contain the SnpMatrix and snp2gene data set}
\usage{
read.ped.to.snpMatrix(
  ped,
  info,
  snp.list = NA,
  ped.header = T,
  ped.sep = "\\t"
)
}
\arguments{
\item{ped}{output of the function \code{from.assay.to.ped}.}

\item{info}{a data frame or a file containing the snp and gene information. The data set contain 4 variables:Chromosome, Genenames, SNPnames, Position.}

\item{snp.list}{a vector contain the SNPs you want to analysis.}

\item{ped.header}{logical. if \code{TRUE}, the ped file's first line should be the colnames of it.}

\item{ped.sep}{a Delimited symbol}
}
\value{
a list contain SnpMatrix and data frame with the row is snps and the col is Chromosome, Genenames, SNPnames, Position:
\item{SnpMatrix}{function output 1: the genotype on each genome loci of each sample}
\item{info}{function output 2: the relationship between the SNPs and genes}
}
\description{
\code{read.ped.to.snpMatrix} read the ped file and output a list contain the SnpMatrix and snp2gene data set.
}
\details{
This function use the output of the function \code{from.assay.to.ped}. 
you should provide the relationship between snps and genes, too.
The output of this function is a list.
}
\examples{
write.table(data.ped, file = "data.with.header.ped", quote = F, sep = "\t", col.names = TRUE, row.names = FALSE)
res <- read.ped.to.snpMatrix(ped = "data.with.header.ped", ped.header = TRUE, info = snp2gene)


}
\author{
Liangjie Liu <liuliangjie@sjtu.edu.cn>
}
