% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\docType{data}
\name{res}
\alias{res}
\title{a dataset of class list for 1536 samples and 92 snps}
\format{
A list with 4 datasets. The variables
are as follows:
\describe{
  \item{genotypes}{A snpmatrix with 1504 rows and 92 columns}
  \item{fam}{The detail information of each sample}
  \item{map}{The detail information of each snp loci}
  \item{info}{The relationship between each snp loci and the genes}
}
}
\source{
unknown.
}
\usage{
res
}
\description{
A dataset containing the information of 1536 samples
and their 92 snp locis data.
}
\keyword{datasets}
