% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.info.R
\docType{data}
\name{snp.info}
\alias{snp.info}
\title{snp.info data for 92 snps}
\format{
A data frame with 92 rows and many variables. The required variables
are as follows:
\describe{
  \item{ABI.ID}{the ID of snp loci match the colnames of assay data}
  \item{Assays.name}{The rs ID of each snp loci}
}
}
\source{
unknown.
}
\usage{
snp.info
}
\description{
A dataset containing the information of 92 snp locis.
}
\keyword{datasets}
