% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp2gene.R
\docType{data}
\name{snp2gene}
\alias{snp2gene}
\title{snp2gene data for 92 snps}
\format{
A data frame with 92 rows and 4 variables. The variables
are as follows:
\describe{
  \item{Chromosome}{The Chromosome}
  \item{Genenames}{The gene name}
  \item{SNPnames}{The rs ID of each snp}
  \item{Position}{The position of each snp}
}
}
\source{
unknown.
}
\usage{
snp2gene
}
\description{
A dataset containing the information of 92 snp locis data including thier position and genes.
}
\keyword{datasets}
