% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggc.example.R
\docType{data}
\name{temp.ggc}
\alias{temp.ggc}
\title{a dataset of class list}
\format{
A list with 7 datasets. The variables
are as follows:
\describe{
  \item{SnpMatrix}{A snpmatrix with 1459 rows and 92 columns}
  \item{gene.info}{A data frame with 92 rows and 4 columns, each row for a snp loci}
  \item{sample.info}{A data frame with 1459 row s and 6 columns, each row for a single sample}
  \item{gg.causal}{the causal relationship between all pairs of selected genes}
  \item{gene.table}{the table of selected genes}
  \item{gg.f}{a matrix for f value}
  \item{method}{the method of reducing dimension}
}
}
\source{
unknown.
}
\usage{
temp.ggc
}
\description{
A example dataset of the output of \code{gene.gene.causal}
}
\keyword{datasets}
