% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiud.example.R
\docType{data}
\name{temp.ggiud}
\alias{temp.ggiud}
\title{a dataset of class list}
\format{
A list with 7 data frame. The variables
are as follows:
\describe{
  \item{SnpMatrix}{A snpmatrix with 1459 rows and 92 columns}
  \item{gene.info}{A data frame with 92 rows and 4 columns, each row for a snp loci}
  \item{sample.info}{A data frame with 1459 row s and 6 columns, each row for a single sample}
  \item{GGI.ud}{the interaction between all pairs of selected genes}
  \item{GENE.table}{the table of selected genes}
  \item{P.matrix}{a matrix for p value}
  \item{S.matrix}{a matrix for s value based on GBIGM}
}
}
\source{
unknown.
}
\usage{
temp.ggiud
}
\description{
A example dataset of the output of \code{gene.gene.interaction}
}
\keyword{datasets}
