% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.after.impute.R
\docType{data}
\name{temp.im}
\alias{temp.im}
\title{a dataset of class list}
\format{
A list with 3 data frame. The variables
are as follows:
\describe{
  \item{SnpMatrix}{A snpmatrix with 1459 rows and 92 columns}
  \item{gene.info}{A data frame with 92 rows and 4 columns, each row for a snp loci}
  \item{sample.info}{A data frame with 1459 row s and 6 columns, each row for a single sample}
}
}
\source{
unknown.
}
\usage{
temp.im
}
\description{
A example dataset of the output of \code{impute.res}
}
\keyword{datasets}
